requirements.txt
google-genai
streamlit
asyncio
aiofiles
pillow
numpy
tqdm
json_repair
anthropic
openai
matplotlib
python-dotenv
pyyaml
google-auth
pydantic-settings
fastapi
python-multipart
uvicorn